refactor(platform): move SQLiteService into global PlatformModule
Phase 6a of the CQRS refactor (#235, first of two commits).
`SQLiteService` is infrastructure — a 135-line better-sqlite3 wrapper
plus a startup schema migration — not a graph concern. Moving it out
of `GraphModule` into the already-@Global `src/platform/platform.module.ts`
(the CqrsModule host from Phase 0) gives infrastructure a dedicated
home and sets up the forwardRef cleanup that lands in the next commit.

Changes:
  - File move: src/modules/graph/sqlite.service.ts
                → src/platform/sqlite.service.ts
    Body unchanged, relative imports adjusted (../../config.js →
    ../config.js, ../../lib/manifest-core.js → ../lib/manifest-core.js)
  - src/platform/platform.module.ts now provides + exports
    SQLiteService alongside CqrsModule. Stays @Global — consumers
    don't need an explicit `imports: [PlatformModule]` entry, which
    simplifies the migration from the proposal doc's non-Global
    approach.
  - src/modules/graph/graph.module.ts drops SQLiteService from
    providers + exports. No new import needed; the @Global platform
    covers it.
  - 8 consumer source files update their relative import path:
    * graph siblings (4): ./sqlite.service.js
         → ../../platform/sqlite.service.js
    * cross-module (4): ../graph/sqlite.service.js
         → ../../platform/sqlite.service.js
  - 4 test files update their relative import path similarly. All
    use Object.create(SQLiteService.prototype) harnesses so the
    change is path-only — no test-body churn.

This commit is architecturally neutral — forwardRef count stays at
7. The cycle-breaking lands in the follow-up commit.

All 561 tests pass; tsc --noEmit clean.
