+++ src/modules/graph/graph.module.ts
@@ -5,16 +5,21 @@ import { GraphController } from "./graph.controller.js";
 import { GraphEventsService } from "./graph-events.service.js";
 import { GraphService } from "./graph.service.js";
 import { GraphWriterService } from "./graph-writer.service.js";
-import { SQLiteService } from "./sqlite.service.js";
 import { WorkItemHsmService } from "./work-item-hsm.service.js";
 import { WorkItemsController } from "./work-items.controller.js";
 import { WorkItemsService } from "./work-items.service.js";
 
+/**
+ * Phase 6a (#235): `SQLiteService` moved to `@Global PlatformModule`
+ * (src/platform/). GraphModule no longer provides or exports it.
+ * Graph's own services still inject `SQLiteService` directly — it's
+ * available via the global platform without an explicit module
+ * import.
+ */
 @Module({
   imports: [],
   controllers: [WorkItemsController, EdgesController, GraphController],
   providers: [
-    SQLiteService,
     WorkItemsService,
     EdgesService,
     GraphService,
@@ -23,7 +28,6 @@ import { WorkItemsService } from "./work-items.service.js";
     WorkItemHsmService,
   ],
   exports: [
-    SQLiteService,
     WorkItemsService,
     EdgesService,
     GraphService,
